# livros-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Livros",
        "description": "Documentação da API de Livros",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/livros": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Lista todos os livros",
                "operationId": "4ec71241f449c1ef2118b675884d8cc9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Filtrar pelo título do livro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "editora",
                        "in": "query",
                        "description": "Filtrar pela editora do livro",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "edicao",
                        "in": "query",
                        "description": "Filtrar pelo número da edição",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ano_publicacao",
                        "in": "query",
                        "description": "Filtrar pelo ano de publicação",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de livros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Livro"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Livros"
                ],
                "summary": "Cadastra um novo livro",
                "operationId": "5a20828782036ae7ef1b08bd8b9b34e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livro"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Livro cadastrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/livros/{id}": {
            "get": {
                "tags": [
                    "Livros"
                ],
                "summary": "Exibe um livro",
                "operationId": "9bd7742c17f0d3e386a8562c01e3e440",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Livros"
                ],
                "summary": "Atualiza um livro",
                "operationId": "13ee1e7476e35757d06ddab15acad152",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Livro"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Livro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Livro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Livros"
                ],
                "summary": "Remove um livro",
                "operationId": "b64c761bf9de3b17940dd9bff70200df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Livro removido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Livro": {
                "title": "Livro",
                "description": "Modelo de Livro",
                "required": [
                    "titulo",
                    "editora",
                    "ano_publicacao",
                    "valor"
                ],
                "properties": {
                    "cod_livro": {
                        "description": "Código do livro",
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "description": "Título do livro",
                        "type": "string",
                        "example": "Dom Casmurro"
                    },
                    "editora": {
                        "description": "Editora do livro",
                        "type": "string",
                        "example": "Editora Globo"
                    },
                    "edicao": {
                        "description": "Número da edição",
                        "type": "integer",
                        "example": 2
                    },
                    "ano_publicacao": {
                        "description": "Ano de publicação",
                        "type": "string",
                        "example": 1899
                    },
                    "valor": {
                        "description": "Valor do livro",
                        "type": "number",
                        "format": "float",
                        "example": 39.9
                    },
                    "autores": {
                        "description": "Lista de autores",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "assuntos": {
                        "description": "Lista de assuntos",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}